Restructure project under src layout
