+++ src/nf_llm/fantasy_football/display.py
